+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fix `evaluate` to return the correct value instead of None for falsy values @thromer
+
 ### Added
 
 ### Changed